Added a basic tick system
There are 3 main speeds in this program:
The max fps (for UI, player actions and the sort that we want to just run quickly)
The tick speed (each game tick will allow for changes in the simulation)
The render speed (every render will draw the game state)
These priorities are organised by time as the render speed can be slower than the tick speed which can be slower than the max fps.

for now, max fps is 64, tps is 8 and renders/second is 16.

The disadvantage of this method is that they need to be divisible by each other otherwise I'd need to use something like time after frame.