Fix for rotation of building plans
After taking the initial coordinates from the building plan, the ox and oy are rotated 90 degrees direction times. This is done by the coordinate rotation formula. May be a bit inefficient, but this is only a process that happens every time a new building is placed and not every tick.