VERY MUCH IN PAIN
Currently the program is not working well. A way to add items on the belts has been added (hold I) and testing of the belts is underway. Unfortunately I am stupid, so it's not quite working yet. Id recommend working on an earlier version until I've finished ironing out these issues. Don't mind the printf statements.