vague updates to item movement
When pressing I it now spawns one of each item. These are colour coded with an array of colours, but the array would be a constant defined at the start if i wasn't lazy and knew it would be gone when actual textures are used. Pressing I outside of a building causes issues. Still having slight issues regarding items skipping spot 3 in buildings occasionally.